Fix bug in function player1MoveReveal of RockPaperScissors
Explicitly check that player2 has already made a move, because we have
to be certain that player2 did not move so that player1 can cancel.
Although the function would have reverted with "invalid move" when it
called into function determineGame. We require it to fail faster (less
gas) and more explicitly.